[feat] pre-release and release do not make any modifications to the repo
This is up to the end user to make the commit during the publishing
phase.
