+++ src/ts-pre-release.ts
@@ -1,9 +1,7 @@
 #!/usr/bin/env node
 import 'colors';
 import * as yargs from 'yargs';
-import * as _ from 'lodash';
 import { exit, run, getConfig, info } from 'ts-publish';
-import { readFileSync, writeFileSync } from 'fs';
 import { normalize } from 'path';
 
 interface IArgs {
@@ -11,24 +9,10 @@ interface IArgs {
 }
 
 const pkg = getConfig('package.json');
-const date = new Date();
 const argv: IArgs = yargs.usage('usage: $0 hook-file')
   .demand(1)
   .help('help')
   .argv;
-
-function changeVersion(dateValue: number): string {
-  const contents = readFileSync('package.json', 'utf8');
-  const lines = contents.split('\n');
-  const newLines = lines.map((line) => {
-    if (_.startsWith(_.trim(line), '"version"')) {
-      return `  "version": "${pkg.version}-beta.${dateValue}",`;
-    }
-    return line;
-  });
-  writeFileSync('package.json', newLines.join('\n'));
-  return `${pkg.version}-beta.${dateValue}`;
-}
 
 function main(): number {
   run('git status -s', (stdout: string) => {
@@ -45,9 +29,6 @@ function main(): number {
     }
   });
 
-  info('CHECKOUT'.cyan, 'build'.green);
-  run('git checkout -b build');
-
   let hook: any;
   const hookPath: string = normalize(`${process.cwd()}/${argv._[0]}`);
   try {
@@ -57,6 +38,9 @@ function main(): number {
     console.log(e.stack);
     throw Error('exit');
   }
+
+  info('CHECKOUT'.cyan, 'build'.green);
+  run('git checkout -b build');
 
   info('HOOK'.cyan, 'running ...');
   try {
@@ -74,13 +58,8 @@ function main(): number {
     }
   });
 
-  info('COMMIT'.cyan);
-  const version = changeVersion(date.valueOf());
-  run('git add package.json -f');
-  run(`git commit -m "[pre-release:${date.valueOf()}]"`);
-
   try {
-    hook.publish('pre-release', version);
+    hook.publish('pre-release', pkg.version);
   } catch (e) {
     info('ERROR'.red, `hook publish error:\n'${e.message}'`);
     console.log(e.stack);
